update code for async
