+++ cassandra-java-driver-3.x-scenario/docker/docker-entrypoint.sh
