+++ cassandra-java-driver-3.x-scenario/docker/startup.sh
