+++ cassandra-java-driver-3.x-scenario/pom.xml
@@ -5,10 +5,10 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.apache.skywalking</groupId>
-    <artifactId>cassandra-3.x-scenario</artifactId>
+    <artifactId>cassandra-java-driver-3.x-scenario</artifactId>
     <version>6.4.0</version>
 
-    <name>skywalking-cassandra-3.x-scenario</name>
+    <name>skywalking-cassandra-java-driver-3.x-scenario</name>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
