+++ cassandra-java-driver-3.x-scenario/src/main/java/test/apache/skywalking/testcase/cassandra/Application.java
