+++ cassandra-java-driver-3.x-scenario/src/main/resources/application.yml
