+++ cassandra-java-driver-3.x-scenario/src/main/resources/log4j2.xml
